♻️ extract trackViewMetrics
